requirements.txt
faiss==1.5.3
faiss-cpu==1.13.2
fastapi==0.129.0
fitz==0.0.1.dev2
gradio==6.6.0
gradio_client==2.0.3
google-genai==1.63.0
langchain==1.2.10
langchain_core==1.2.13
langchain_google_genai==4.2.0
langchain_text_splitters==1.1.1
langserve==0.3.3
numpy==2.4.2
protobuf==6.33.5
pydantic==2.12.5
python==3.11.14
python-dotenv==1.2.1
python_docx==1.2.0
PyMuPDF==1.27.1
Requests==2.32.5
uvicorn==0.41.0